Move Python from src/mit_core to src
It’s not a real package, it’s just a bunch of modules. Simplify the source
code layout, as we do in src/specializer/ and src/features/.
