+++ python/gen-trap-enums
@@ -8,9 +8,9 @@
 # THIS PROGRAM IS PROVIDED AS IS, WITH NO WARRANTY. USE IS AT THE USER’S
 # RISK.
 
-from mit_core.code_util import Code, enum_to_python, copyright_banner
-import mit_core.stack
-mit_core.stack.TYPE_SIZE_UNKNOWN = 0
+from code_util import Code, enum_to_python, copyright_banner
+import stack
+stack.TYPE_SIZE_UNKNOWN = 0
 
 import traps
