+++ src/features/collate-type-sizes
@@ -12,8 +12,8 @@ import sys
 import re
 from pprint import pformat
 
-import mit_core.stack
-mit_core.stack.TYPE_SIZE_UNKNOWN = 0
+import stack
+stack.TYPE_SIZE_UNKNOWN = 0
 
 from traps import *
