+++ src/features/gen-trap-codes
@@ -8,9 +8,9 @@
 # THIS PROGRAM IS PROVIDED AS IS, WITH NO WARRANTY. USE IS AT THE USER’S
 # RISK.
 
-from mit_core.code_util import Code, copyright_banner, enum_to_c
-import mit_core.stack
-mit_core.stack.TYPE_SIZE_UNKNOWN = 0
+from code_util import Code, copyright_banner, enum_to_c
+import stack
+stack.TYPE_SIZE_UNKNOWN = 0
 
 from traps import *
 from trap_errors import TrapErrorCode
