+++ src/features/gen-traps
@@ -8,15 +8,15 @@
 # THIS PROGRAM IS PROVIDED AS IS, WITH NO WARRANTY. USE IS AT THE USER’S
 # RISK.
 
-from mit_core.code_util import Code, copyright_banner
-from mit_core.instruction_gen import *
-from mit_core.params import word_bytes
+from code_util import Code, copyright_banner
+from instruction_gen import *
+from params import word_bytes
 
 
 # Get type sizes and inject them into instruction_gen
 from type_sizes import type_sizes
-import mit_core.stack
-mit_core.stack.type_wordses.update(
+import stack
+stack.type_wordses.update(
     {type: (size + word_bytes - 1) // word_bytes
      for type, size in type_sizes.items()}
 )
