+++ src/features/gen-type-sizes
@@ -10,9 +10,9 @@
 
 import re
 
-from mit_core.code_util import Code
-import mit_core.stack
-mit_core.stack.TYPE_SIZE_UNKNOWN = 0
+from code_util import Code
+import stack
+stack.TYPE_SIZE_UNKNOWN = 0
 
 from traps import LibInstruction
