+++ src/gen-main
@@ -8,9 +8,8 @@
 # THIS PROGRAM IS PROVIDED AS IS, WITH NO WARRANTY. USE IS AT THE USER’S
 # RISK.
 
-from mit_core.code_util import Code, copyright_banner, disable_warnings
-from mit_core.spec import MitErrorCode
-
+from code_util import Code, copyright_banner, disable_warnings
+from spec import MitErrorCode
 from features.trap_errors import TrapErrorCode
