+++ src/gen-opcodes
@@ -8,8 +8,8 @@
 # THIS PROGRAM IS PROVIDED AS IS, WITH NO WARRANTY. USE IS AT THE USER’S
 # RISK.
 
-from mit_core.spec import Instruction, ExtraInstruction, MitErrorCode
-from mit_core.code_util import Code, copyright_banner, enum_to_c
+from spec import Instruction, ExtraInstruction, MitErrorCode
+from code_util import Code, copyright_banner, enum_to_c
 
 
 GENERATOR_PROGRAM = 'gen-opcodes'
