+++ src/gen-registers
@@ -8,8 +8,8 @@
 # THIS PROGRAM IS PROVIDED AS IS, WITH NO WARRANTY. USE IS AT THE USER’S
 # RISK.
 
-from mit_core.code_util import Code, copyright_banner, unrestrict
-from mit_core.spec import Register
+from code_util import Code, copyright_banner, unrestrict
+from spec import Register
 
 
 GENERATOR_PROGRAM = 'gen-registers'
