+++ src/spec.py
@@ -10,12 +10,12 @@
 import os
 from enum import Enum, IntEnum, unique
 
-from .autonumber import AutoNumber
-from .params import word_bytes
-from .code_util import Code
-from .instruction import InstructionEnum
-from .instruction_gen import dispatch
-from .stack import StackEffect, pop_stack, push_stack
+from autonumber import AutoNumber
+from params import word_bytes
+from code_util import Code
+from instruction import InstructionEnum
+from instruction_gen import dispatch
+from stack import StackEffect, pop_stack, push_stack
 
 
 # Global constants (see module params for build-time parameters)
