+++ src/specializer/specializer.am
@@ -83,10 +83,10 @@ specialize-once: mit@PACKAGE_SUFFIX@$(EXEEXT)
 	cd $(abs_builddir) && \
 	$(MAKE)
 
-%D%/labels.json: mit_core/code_util.py mit_core/params.py mit_core/instruction.py mit_core/spec.py %D%/profile.json %D%/simulate-jit %D%/path.py %D%/profile.py
+%D%/labels.json: code_util.py params.py instruction.py spec.py %D%/profile.json %D%/simulate-jit %D%/path.py %D%/profile.py
 	$(PYTHON_WITH_PATH) $(srcdir)/%D%/simulate-jit $(srcdir)/%D%/profile.json $@
 
-%D%/specializer.c: mit_core/code_util.py mit_core/instruction.py mit_core/stack.py mit_core/spec.py mit_core/instruction_gen.py %D%/specializer.am %D%/gen-specializer %D%/specializer.py %D%/spec.py %D%/labels.json
+%D%/specializer.c: code_util.py instruction.py stack.py spec.py instruction_gen.py %D%/specializer.am %D%/gen-specializer %D%/specializer.py %D%/specializer_spec.py %D%/labels.json
 	$(PYTHON_WITH_PATH) $(srcdir)/%D%/gen-specializer %D%/labels.json > %D%/specializer.c || ( rm -f %D%/specializer.c; exit 1 )
 
 %D%/specializer.lo: include/mit/mit.h
@@ -99,7 +99,7 @@ specialize: mit@PACKAGE_SUFFIX@$(EXEEXT)
 EXTRA_DIST += \
 	%D%/path.py \
 	%D%/profile.py \
-	%D%/spec.py \
+	%D%/specializer_spec.py \
 	%D%/specializer.py \
 	%D%/simulate-jit \
 	%D%/gen-specializer \
